+++ internal/authz/authz.go
@@ -110,6 +110,8 @@ type AuthorizerInterface interface {
 	AuthorizeListStores(ctx context.Context) error
 	ListAuthorizedStores(ctx context.Context) ([]string, error)
 	GetModulesForWriteRequest(req *openfgav1.WriteRequest, typesys *typesystem.TypeSystem) ([]string, error)
+	AccessControlStoreID() string
+	IsNoop() bool
 }
 
 type NoopAuthorizer struct{}
@@ -136,6 +138,14 @@ func (a *NoopAuthorizer) ListAuthorizedStores(ctx context.Context) ([]string, er
 
 func (a *NoopAuthorizer) GetModulesForWriteRequest(req *openfgav1.WriteRequest, typesys *typesystem.TypeSystem) ([]string, error) {
 	return nil, nil
+}
+
+func (a *NoopAuthorizer) AccessControlStoreID() string {
+	return ""
+}
+
+func (a *NoopAuthorizer) IsNoop() bool {
+	return true
 }
 
 type Authorizer struct {
@@ -201,6 +211,17 @@ func (a *Authorizer) getRelation(apiMethod string) (string, error) {
 	default:
 		return "", AuthorizationError{Err: ErrUnknownAPIMethod}.Err
 	}
+}
+
+func (a *Authorizer) AccessControlStoreID() string {
+	if a.config != nil {
+		return a.config.StoreID
+	}
+	return ""
+}
+
+func (a *Authorizer) IsNoop() bool {
+	return false
 }
 
 // Authorize checks if the user has access to the resource.