+++ pkg/storage/sqlite/sqlite.go
@@ -9,8 +9,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/openfga/openfga/pkg/encoder"
-
 	sq "github.com/Masterminds/squirrel"
 	"github.com/oklog/ulid/v2"
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
@@ -44,7 +42,6 @@ type Datastore struct {
 	stbl                   sq.StatementBuilderType
 	db                     *sql.DB
 	logger                 logger.Logger
-	tokenSerializer        encoder.ContinuationTokenSerializer
 	dbStatsCollector       prometheus.Collector
 	maxTuplesPerWriteField int
 	maxTypesPerModelField  int
@@ -115,17 +112,12 @@ func New(uri string, cfg *sqlcommon.Config) (*Datastore, error) {
 		}
 	}
 
-	if cfg.TokenSerializer == nil {
-		cfg.TokenSerializer = sqlcommon.NewSQLContinuationTokenSerializer()
-	}
-
 	stbl := sq.StatementBuilder.RunWith(db)
 
 	return &Datastore{
 		stbl:                   stbl,
 		db:                     db,
 		logger:                 cfg.Logger,
-		tokenSerializer:        cfg.TokenSerializer,
 		dbStatsCollector:       collector,
 		maxTuplesPerWriteField: cfg.MaxTuplesPerWriteField,
 		maxTypesPerModelField:  cfg.MaxTypesPerModelField,
@@ -154,18 +146,13 @@ func (s *Datastore) Read(
 }
 
 // ReadPage see [storage.RelationshipTupleReader].ReadPage.
-func (s *Datastore) ReadPage(
-	ctx context.Context,
-	store string,
-	tupleKey *openfgav1.TupleKey,
-	options storage.ReadPageOptions,
-) ([]*openfgav1.Tuple, []byte, error) {
+func (s *Datastore) ReadPage(ctx context.Context, store string, tupleKey *openfgav1.TupleKey, options storage.ReadPageOptions) ([]*openfgav1.Tuple, string, error) {
 	ctx, span := startTrace(ctx, "ReadPage")
 	defer span.End()
 
 	iter, err := s.read(ctx, store, tupleKey, &options)
 	if err != nil {
-		return nil, nil, err
+		return nil, "", err
 	}
 	defer iter.Stop()
 
@@ -207,11 +194,8 @@ func (s *Datastore) read(ctx context.Context, store string, tupleKey *openfgav1.
 		})
 	}
 	if options != nil && options.Pagination.From != "" {
-		token, err := sqlcommon.UnmarshallContToken(options.Pagination.From)
-		if err != nil {
-			return nil, err
-		}
-		sb = sb.Where(sq.GtOrEq{"ulid": token.Ulid})
+		token := options.Pagination.From
+		sb = sb.Where(sq.GtOrEq{"ulid": token})
 	}
 	if options != nil && options.Pagination.PageSize != 0 {
 		sb = sb.Limit(uint64(options.Pagination.PageSize + 1)) // + 1 is used to determine whether to return a continuation token.
@@ -642,11 +626,7 @@ func (s *Datastore) ReadAuthorizationModel(ctx context.Context, store string, mo
 }
 
 // ReadAuthorizationModels see [storage.AuthorizationModelReadBackend].ReadAuthorizationModels.
-func (s *Datastore) ReadAuthorizationModels(
-	ctx context.Context,
-	store string,
-	options storage.ReadAuthorizationModelsOptions,
-) ([]*openfgav1.AuthorizationModel, []byte, error) {
+func (s *Datastore) ReadAuthorizationModels(ctx context.Context, store string, options storage.ReadAuthorizationModelsOptions) ([]*openfgav1.AuthorizationModel, string, error) {
 	ctx, span := startTrace(ctx, "ReadAuthorizationModels")
 	defer span.End()
 
@@ -657,11 +637,7 @@ func (s *Datastore) ReadAuthorizationModels(
 		OrderBy("authorization_model_id desc")
 
 	if options.Pagination.From != "" {
-		token, _, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, nil, err
-		}
-		sb = sb.Where(sq.LtOrEq{"authorization_model_id": token})
+		sb = sb.Where(sq.LtOrEq{"authorization_model_id": options.Pagination.From})
 	}
 	if options.Pagination.PageSize > 0 {
 		sb = sb.Limit(uint64(options.Pagination.PageSize + 1)) // + 1 is used to determine whether to return a continuation token.
@@ -669,7 +645,7 @@ func (s *Datastore) ReadAuthorizationModels(
 
 	rows, err := sb.QueryContext(ctx)
 	if err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 	defer rows.Close()
 
@@ -678,33 +654,28 @@ func (s *Datastore) ReadAuthorizationModels(
 	var marshalledModel []byte
 
 	models := make([]*openfgav1.AuthorizationModel, 0, options.Pagination.PageSize)
-	var token []byte
+	var token string
 
 	for rows.Next() {
 		err = rows.Scan(&modelID, &schemaVersion, &marshalledModel)
 		if err != nil {
-			return nil, nil, HandleSQLError(err)
+			return nil, "", HandleSQLError(err)
 		}
 
 		if options.Pagination.PageSize > 0 && len(models) >= options.Pagination.PageSize {
-			token, err = s.tokenSerializer.Serialize(modelID, "")
-			if err != nil {
-				return nil, nil, err
-			}
-
-			return models, token, nil
+			return models, modelID, nil
 		}
 
 		var model openfgav1.AuthorizationModel
 		if err := proto.Unmarshal(marshalledModel, &model); err != nil {
-			return nil, nil, err
+			return nil, "", err
 		}
 
 		models = append(models, &model)
 	}
 
 	if err := rows.Err(); err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 
 	return models, token, nil
@@ -833,7 +804,7 @@ func (s *Datastore) GetStore(ctx context.Context, id string) (*openfgav1.Store, 
 }
 
 // ListStores provides a paginated list of all stores present in the storage.
-func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOptions) ([]*openfgav1.Store, []byte, error) {
+func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOptions) ([]*openfgav1.Store, string, error) {
 	ctx, span := startTrace(ctx, "ListStores")
 	defer span.End()
 
@@ -854,11 +825,7 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 		OrderBy("id")
 
 	if options.Pagination.From != "" {
-		token, _, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, nil, err
-		}
-		sb = sb.Where(sq.GtOrEq{"id": token})
+		sb = sb.Where(sq.GtOrEq{"id": options.Pagination.From})
 	}
 	if options.Pagination.PageSize > 0 {
 		sb = sb.Limit(uint64(options.Pagination.PageSize + 1)) // + 1 is used to determine whether to return a continuation token.
@@ -866,7 +833,7 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 
 	rows, err := sb.QueryContext(ctx)
 	if err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 	defer rows.Close()
 
@@ -877,7 +844,7 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 		var createdAt, updatedAt time.Time
 		err := rows.Scan(&id, &name, &createdAt, &updatedAt)
 		if err != nil {
-			return nil, nil, HandleSQLError(err)
+			return nil, "", HandleSQLError(err)
 		}
 
 		stores = append(stores, &openfgav1.Store{
@@ -889,18 +856,14 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 	}
 
 	if err := rows.Err(); err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 
 	if len(stores) > options.Pagination.PageSize {
-		contToken, err := s.tokenSerializer.Serialize(id, "")
-		if err != nil {
-			return nil, nil, err
-		}
-		return stores[:options.Pagination.PageSize], contToken, nil
+		return stores[:options.Pagination.PageSize], id, nil
 	}
 
-	return stores, nil, nil
+	return stores, "", nil
 }
 
 // DeleteStore removes a store from storage.
@@ -978,12 +941,7 @@ func (s *Datastore) ReadAssertions(ctx context.Context, store, modelID string) (
 }
 
 // ReadChanges see [storage.ChangelogBackend].ReadChanges.
-func (s *Datastore) ReadChanges(
-	ctx context.Context,
-	store string,
-	filter storage.ReadChangesFilter,
-	options storage.ReadChangesOptions,
-) ([]*openfgav1.TupleChange, []byte, error) {
+func (s *Datastore) ReadChanges(ctx context.Context, store string, filter storage.ReadChangesFilter, options storage.ReadChangesOptions) ([]*openfgav1.TupleChange, string, error) {
 	ctx, span := startTrace(ctx, "ReadChanges")
 	defer span.End()
 
@@ -1011,15 +969,7 @@ func (s *Datastore) ReadChanges(
 		sb = sb.Where(sq.Eq{"object_type": objectTypeFilter})
 	}
 	if options.Pagination.From != "" {
-		token, objectType, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, nil, err
-		}
-		if objectType != objectTypeFilter {
-			return nil, nil, storage.ErrMismatchObjectType
-		}
-
-		sb = sqlcommon.AddFromUlid(sb, token, options.SortDesc)
+		sb = sqlcommon.AddFromUlid(sb, options.Pagination.From, options.SortDesc)
 	}
 	if options.Pagination.PageSize > 0 {
 		sb = sb.Limit(uint64(options.Pagination.PageSize)) // + 1 is NOT used here as we always return a continuation token.
@@ -1027,7 +977,7 @@ func (s *Datastore) ReadChanges(
 
 	rows, err := sb.QueryContext(ctx)
 	if err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 	defer rows.Close()
 
@@ -1054,14 +1004,14 @@ func (s *Datastore) ReadChanges(
 			&insertedAt,
 		)
 		if err != nil {
-			return nil, nil, HandleSQLError(err)
+			return nil, "", HandleSQLError(err)
 		}
 
 		var conditionContextStruct structpb.Struct
 		if conditionName.String != "" {
 			if conditionContext != nil {
 				if err := proto.Unmarshal(conditionContext, &conditionContextStruct); err != nil {
-					return nil, nil, err
+					return nil, "", err
 				}
 			}
 		}
@@ -1082,15 +1032,10 @@ func (s *Datastore) ReadChanges(
 	}
 
 	if len(changes) == 0 {
-		return nil, nil, storage.ErrNotFound
+		return nil, "", storage.ErrNotFound
 	}
 
-	contToken, err := s.tokenSerializer.Serialize(ulid, objectTypeFilter)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return changes, contToken, nil
+	return changes, ulid, nil
 }
 
 // IsReady see [sqlcommon.IsReady].